% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/start_project.R
\name{start_project}
\alias{start_project}
\title{Start a new project with the varsityblues templates}
\usage{
start_project(path = ".")
}
\arguments{
\item{path}{Path to the new project (defaults to working directory)}
}
\description{
Start a new project with the varsityblues templates
}
