% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varsityblues.R
\docType{package}
\name{varsityblues}
\alias{varsityblues}
\alias{varsityblues-package}
\title{varsityblues: Writing your UofT assignmente using R Markdown}
\description{
Using the rmarkdown package and LaTeX, this package aims to 
    encourage reproducible research and better reproducibility practices using 
    simple Markdown syntax while embedding all of the R code to produce plots
    and analyses as well. Varsityblues specifically works for the University of 
    Toronto assignments format.
}
\author{
\strong{Maintainer}: Mauricio Vargas \email{m.vargas.sepulveda@mail.utoronto.ca} (\href{https://orcid.org/0000-0003-1017-7574}{ORCID}) [copyright holder]

}
